fix: sort channel recurring reminders among recurrent reminders
Before, channel recurring reminders where categorized as channel
reminders, which gave them a "Complete" action.

But it doesn't make sense to complete a recurring reminder – and it even
causes the reminder not to be scheduled again, and eventually to stop
working.

This commit, by categorizing channel recurring reminders as recurring
reminders, make sure they won't be marked as "Completed" by mistake.

Probably closes #167